backend/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
ffmpeg-python
requests
psycopg2-binary
supabase
python-dotenv
httpx
